meged ase2HDF5 and MDA2HDF% into ToHDF5
